add eslint and cleanup a few things.
