+++ src/App.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import HomePage from './homepage';
 
-
 export default function App() {
-    return (
-        <HomePage />
-    );
+  return (
+    <HomePage />
+  );
 }
